[API_PARSER][CYBEREASON] Refacto & split into smaller files + get bulk of max 1000 logs to avoid API error (#456)
 ### Updated
- **[API_PARSER][CYBEREASON]** Refacto & split into smaller files + get bulk of max 1000 logs to avoid API error

---------
