+++ CHANGELOG
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - [MONITOR] Avoid wrong node state during monitoring operations
+### Changed
+- [API_PARSER] [CYBEREASON] Refacto code and avoid API errors by getting bulks of 1000 logs
 
 
 ## [2.17.0] - 2024-11-25
